# app.py
from flask import Flask, request, jsonify
from flask_swagger_ui import get_swaggerui_blueprint
from flask_cors import CORS
import sqlite3
import os
from pathlib import Path

app = Flask(__name__)
CORS(app)
DB = os.getenv("DB_PATH", "data/todos.db")

# Configuration Swagger
SWAGGER_URL = '/api/docs'
API_URL = '/static/swagger.json'

swaggerui_blueprint = get_swaggerui_blueprint(
    SWAGGER_URL,
    API_URL,
    config={
        'app_name': "Todo API"
    }
)

app.register_blueprint(swaggerui_blueprint, url_prefix=SWAGGER_URL)


def ensure_db_directory():
    Path(DB).parent.mkdir(parents=True, exist_ok=True)


def get_connection():
    ensure_db_directory()
    conn = sqlite3.connect(DB)
    conn.execute("CREATE TABLE IF NOT EXISTS todos (id INTEGER PRIMARY KEY, task TEXT)")
    return conn

def init_db():
    conn = get_connection()
    conn.commit()
    conn.close()

@app.route("/api/v1/todos", methods=["GET"])
def get_todos():
    conn = get_connection()
    c = conn.cursor()
    c.execute("SELECT * FROM todos")
    rows = c.fetchall()
    conn.close()

    todos = [{"id": r[0], "task": r[1]} for r in rows]
    return jsonify(todos)

@app.route("/api/v1/todos", methods=["POST"])
def add_todo():
    data = request.get_json(silent=True) or {}
    task = str(data.get("task", "")).strip()
    if not task:
        return jsonify({"error": "Le champ 'task' est obligatoire"}), 400

    conn = get_connection()
    c = conn.cursor()
    c.execute("INSERT INTO todos (task) VALUES (?)", (task,))
    conn.commit()
    conn.close()

    return jsonify({"message": "Tâche ajoutée"}), 201

@app.route("/api/v1/todos/<int:id>", methods=["PUT"])
def update_todo(id):
    data = request.get_json(silent=True) or {}
    task = str(data.get("task", "")).strip()
    if not task:
        return jsonify({"error": "Le champ 'task' est obligatoire"}), 400

    conn = get_connection()
    c = conn.cursor()
    c.execute("UPDATE todos SET task=? WHERE id=?", (task, id))
    conn.commit()
    conn.close()

    return jsonify({"message": "Tâche mise à jour"})

@app.route("/api/v1/todos/<int:id>", methods=["DELETE"])
def delete_todo(id):
    conn = get_connection()
    c = conn.cursor()
    c.execute("DELETE FROM todos WHERE id=?", (id,))
    conn.commit()
    conn.close()

    return jsonify({"message": "Tâche supprimée"})


@app.route("/health", methods=["GET"])
def health_check():
    conn = get_connection()
    c = conn.cursor()
    c.execute("SELECT 1")
    conn.close()
    return jsonify({"status": "ok"})

if __name__ == "__main__":
    init_db()
    app.run(host="0.0.0.0", port=int(os.getenv("PORT", "9090")))